Move all SMI functionality into MAC module
Add some docs and make LAN87XXA phy implmementation a bit more clear
