+++ Cargo.toml
@@ -15,7 +15,7 @@ travis-ci = { repository = "astro/stm32-eth", branch = "master" }
 maintenance = { status = "experimental" }
 
 [package.metadata.docs.rs]
-features = ["smoltcp-phy", "stm32f429", "smoltcp/socket-tcp"]
+features = ["smoltcp-phy", "smoltcp/socket-raw", "stm32f429"]
 
 [dependencies]
 volatile-register = "0.2"
